+++ src/main/java/de/samply/frontend/dto/FormField.java
@@ -19,6 +19,7 @@ public record FormField(
         String labelDisplayName,
         String labelDescription,
         FormFieldGroup[] groups,
+        String[] properties,
         DataType type,
         FormFieldValue[] allowedValues,
         Boolean mandatory,