+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## [1.0.0 - 2026-07-09]
+## [1.0.0 - 2026-07-10]
 ### Added
 - First version of the project
 - Spring Application
@@ -265,6 +265,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Form field blocks
 - Delete Project API endpoint
 - Delete Form Field Block API endpoint
+- Form field properties
 
 
 ### Changed